refine catalyst wiki deletion

+++ includes.php
@@ -475,7 +475,8 @@ function delete_wiki( string $wiki, string $serverUri = null ): ?string {
 		return 'Wiki already deleted.';
 	}
 
-	if ( $wikiData['backend'] == 'catalyst' ) {
+	if ( $wikiData['backend'] == 'catalyst' && $wikiData['catalystId'] != null ) {
+		//FIXME: handle error; don't print wiki deleted if error occurs
 		$catalystApi->deleteEnvironment( $wikiData['catalystId'] );
 	} else {
 		$error = shell_echo( __DIR__ . '/deletewiki.sh',
